+++ R/ent_similarity.R
@@ -21,11 +21,8 @@
 #' `similarities` matrix: species are assumed to be in the same order.
 #' @param ... Unused.
 #'
-#' @return A tibble with the site names, the estimators used and the estimated entropy.
+#' @returns A tibble with the site names, the estimators used and the estimated entropy.
 #'
-#' @references
-#' \insertAllCited{}
-#' 
 #' @examples
 #' # Similarity matrix
 #' Z <- fun_similarity(paracou_6_fundist)
@@ -429,7 +426,7 @@ ent_similarity.species_distribution <- function(
 #' @param w_v A weight.
 #' @param p_V_Ns An intermediate computation.
 #'
-#' @return A number.
+#' @returns A number.
 #' @noRd
 S_v <- function(
     species_index,