+++ R/package.R
@@ -12,11 +12,80 @@
 "_PACKAGE"
 
 
-# Functions to reexport
+# Functions to reexport ----
 #' @export
 ggplot2::autoplot
 base::as.numeric
 base::as.matrix
+
+#' ggplot method to plot wmppp objects
+#' 
+#' This method is from the dbmss package. See [dbmss::autoplot.wmppp].
+#'
+#' @param object an object to be plotted.
+#' @param ... extra arguments, currently unused.
+#' @param show.window if `TRUE`, the borders of the window containing the points are shown on the point map.
+#' @param MaxPointTypes the maximum number of different point types to show. 
+#' If the point set contains more of them, the less frequent ones are gathered as "Other". 
+#' This number must be limited for readability and not to exceed the number of colors offered by the palette.
+#' @param Other the name of the point types gathered as "Other"
+#' @param main the title of the plot.
+#' @param xlab the X-axis label.
+#' @param ylab the Y-axis label.
+#' @param LegendLabels a vector of characters. 
+#' The first two items describe the observed and null-hypothesis curves, the third and last item the confidence interval.
+#' To be used only in plots with two curves (typically observed and expected values).
+#' The default is `NULL` to display the full description of functions.
+#' @param labelSize the guide of the point size legend in point maps, i.e. what the `PointSize` mark represents. 
+#' @param labelColor the guide of the point color legend in point maps, i.e. what the `PointType` mark represents. 
+#' @param palette The color palette used to display point types in maps. See [ggplot2::scale_colour_brewer].
+#' @param windowColor the color used to draw the limits of the windows in point maps. 
+#' @param windowFill the color used to fill the windows in point maps.
+#' @param alpha the opacity of the confidence envelope (in function values) or the points (in maps), between 0 and 1.
+#'
+#' @returns A [ggplot2::ggplot].
+#' @export
+#'
+#' @examples
+#' autoplot(paracou_6_wmppp)
+#' 
+autoplot.wmppp <- function(
+    object, 
+    ..., 
+    show.window = TRUE, 
+    MaxPointTypes = 6, 
+    Other = "Other",
+    main = NULL, 
+    xlab = NULL, 
+    ylab = NULL, 
+    LegendLabels = NULL, 
+    labelSize = "Weight", 
+    labelColor = "Type", 
+    palette="Set1",
+    windowColor = "black", 
+    windowFill = "transparent", 
+    alpha = 1) {
+  
+  return(
+    dbmss:::autoplot.wmppp(
+      object = object, 
+      ..., 
+      show.window = show.window, 
+      MaxPointTypes = 6, 
+      Other = Other,
+      main = main, 
+      xlab = xlab, 
+      ylab = ylab, 
+      LegendLabels = LegendLabels, 
+      labelSize = labelSize, 
+      labelColor = labelColor, 
+      palette = palette,
+      windowColor = windowColor, 
+      windowFill = windowFill, 
+      alpha = alpha
+    )
+  )
+}
 
 
 #  Initialization ----
@@ -93,6 +162,8 @@ non_species_columns <- c(
 #' @format A matrix.
 #' @source Permanent data census of Paracou: <https://paracou.cirad.fr/>
 #' @seealso [paracou_6_abd], [paracou_6_taxo]
+#' @references
+#' \insertAllCited{}
 #' 
 "paracou_6_fundist"
 
@@ -149,7 +220,10 @@ non_species_columns <- c(
 #' 
 #' @param abd A vector of positive integers.
 #'
-#' @return The value of A.
+#' @returns The value of A.
+#' @references
+#' \insertAllCited{}
+#' 
 #' @noRd
 #'
 chao_A <- function(abd) {
@@ -177,6 +251,47 @@ chao_A <- function(abd) {
 }
 
 
+
+
+#' as_named_vector.character
+#' 
+#' Counts the number of points of a `character` vector and returns a named vector.
+#' Names are the items of the character vector. 
+#' This is equivalent to `as.numeric(table(x))` but `table()`
+#' looses the names.
+#'
+#' @param x a character vector.
+#'
+#' @returns A named vector with the number of items by name.
+#' @noRd
+#'
+as_named_vector.character <- function(x){
+  # Count the number of items. Returns a 1D array, not a vector.
+  the_array <- tapply(x, x, length)
+  the_vector <- as.vector(the_array)
+  # Add the names
+  names(the_vector) <- names(the_array) 
+  return(the_vector)
+}
+
+
+#' as_named_vector.wmppp
+#' 
+#' Counts the number of points of a `wmppp` object and returns a named vector.
+#' Names are the point types. 
+#' This is equivalent to `as.numeric(table(X$marks$PointType))` but `table()`
+#' looses the names.
+#'
+#' @param X a [dbmss::wmppp] object, i.e. a weighted, marked planar point pattern.
+#'
+#' @returns A named vector with the number of points by type.
+#' @noRd
+#'
+as_named_vector.wmppp <- function(X){
+  # Count the number of points by type
+  return(as_named_vector.character(X$marks$PointType))
+}
+
 #' check_divent_args
 #'
 #' Checks the arguments of a function of the package divent
@@ -188,47 +303,56 @@ chao_A <- function(abd) {
 #' The function is always called without arguments.
 #' Its arguments exist only for documentation.
 #' 
-#' @param abundances An object of class [abundances].
-#' @param alpha The risk level, 5% by default.
-#' @param as_numeric If `TRUE`, a number or a numeric vector is returned rather than a tibble.
-#' @param check_arguments If `TRUE`, the function arguments are verified.
+#' @param abundances an object of class [abundances].
+#' @param alpha the risk level, 5% by default.
+#' @param as_numeric if `TRUE`, a number or a numeric vector is returned rather than a tibble.
+#' @param check_arguments if `TRUE`, the function arguments are verified.
 #' Should be set to `FALSE` to save time when the arguments have been checked elsewhere.
-#' @param coverage_estimator An estimator of sample coverage used by [coverage].
-#' @param distances A distance matrix or an object of class [stats::dist]
-#' @param estimator An estimator of asymptotic entropy, diversity or richness.
-#' @param gamma If `TRUE`, \eqn{\gamma} diversity, i.e. diversity of the metacommunity, is computed.
-#' @param jack_alpha The risk level, 5% by default, used to optimize the jackknife order.
-#' @param jack_max The highest jackknife order allowed. Default is 10.
-#' @param k The order of Hurlbert's diversity.
-#' @param level The level of interpolation or extrapolation. 
+#' @param correction the edge-effect correction to apply when estimating
+#' the number of neighbors or the K function with [spatstat.explore::Kest].
+#' Default is "isotropic".
+#' @param coverage_estimator an estimator of sample coverage used by [coverage].
+#' @param distances a distance matrix or an object of class [stats::dist]
+#' @param estimator an estimator of asymptotic entropy, diversity or richness.
+#' @param gamma if `TRUE`, \eqn{\gamma} diversity, i.e. diversity of the metacommunity, is computed.
+#' @param global if `TRUE`, a global envelope sensu \insertCite{Duranton2005}{divent} is calculated.
+#' @param jack_alpha the risk level, 5% by default, used to optimize the jackknife order.
+#' @param jack_max the highest jackknife order allowed. Default is 10.
+#' @param k the order of Hurlbert's diversity.
+#' @param level the level of interpolation or extrapolation. 
 #' It may be a sample size (an integer) or a sample coverage 
 #' (a number between 0 and 1).
 #' If not `NULL`, the asymptotic `estimator` is ignored.
-#' @param n_simulations The number of simulations used to estimate the confidence envelope.
-#' @param normalize If `TRUE`, phylogenetic is normalized: the height of the tree is set to 1.
-#' @param probability_estimator A string containing one of the possible estimators
+#' @param n_simulations the number of simulations used to estimate the confidence envelope.
+#' @param normalize if `TRUE`, phylogenetic is normalized: the height of the tree is set to 1.
+#' @param probability_estimator a string containing one of the possible estimators
 #' of the probability distribution (see [probabilities]). 
 #' Used only for extrapolation.
-#' @param q The order of diversity.
-#' @param rate The decay rate of the exponential similarity.
-#' @param richness_estimator An estimator of richness to evaluate the total number of species,
-#' see [div_richness]. Used for interpolation and extrapolation.
-#' @param sample_coverage The sample coverage of `x` calculated elsewhere. 
+#' @param q the order of diversity.
+#' @param r a vector of distances. 
+#' If `NULL` accumulation is along `n`, 
+#' else neighbors are accumulated in circles of radius `r`.
+#' @param rate the decay rate of the exponential similarity.
+#' @param richness_estimator an estimator of richness to evaluate the total number of species,
+#' see [div_richness]. used for interpolation and extrapolation.
+#' @param sample_coverage the sample coverage of `x` calculated elsewhere. 
 #' Used to calculate the gamma diversity of meta-communities, see details. 
-#' @param show_progress If TRUE, a progress bar is shown during long computations. 
-#' @param similarities A similarity matrix, that can be obtained by [fun_similarity].
+#' @param show_progress if TRUE, a progress bar is shown during long computations. 
+#' @param similarities a similarity matrix, that can be obtained by [fun_similarity].
 #' Its default value is the identity matrix.
-#' @param species_distribution An object of class [species_distribution].
-#' @param tree An ultrametric, phylogenetic tree.
+#' @param species_distribution an object of class [species_distribution].
+#' @param tree an ultrametric, phylogenetic tree.
 #' May be an object of class [phylo_divent], [ape::phylo], [ade4::phylog] or [stats::hclust]. 
-#' @param unveiling A string containing one of the possible unveiling methods 
+#' @param unveiling a string containing one of the possible unveiling methods 
 #' to estimate the probabilities of the unobserved species (see [probabilities]).
 #' Used only for extrapolation.
 #' @param use.names if `TRUE`, the names of the `species_distribution` are kept 
 #' in the matrix or vector they are converted to.
-#' @param weights The weights of the sites of the species distributions.
+#' @param weights the weights of the sites of the species distributions.
+#' @param X a spatialized community 
+#' (A [dbmss::wmppp] object with `PointType` values as species names.)
 #'
-#' @return Returns `TRUE` or stops if a problem is detected.
+#' @returns Returns `TRUE` or stops if a problem is detected.
 #' 
 #' @export
 #'
@@ -239,10 +363,12 @@ check_divent_args <- function(
     alpha = NULL,
     as_numeric = NULL,
     check_arguments = NULL,
+    correction = NULL,
     coverage_estimator = NULL,
     distances = NULL,
     estimator = NULL,
     gamma = NULL,
+    global = NULL,
     jack_alpha = NULL,
     jack_max = NULL,
     k = NULL,
@@ -251,6 +377,7 @@ check_divent_args <- function(
     normalize = NULL,
     probability_estimator = NULL,
     q = NULL,
+    r = NULL,
     rate = NULL,
     richness_estimator = NULL,
     sample_coverage = NULL,
@@ -260,11 +387,12 @@ check_divent_args <- function(
     tree = NULL,
     use.names = NULL,
     unveiling = NULL,
-    weights = NULL) {
+    weights = NULL,
+    X = NULL) {
 
   # Verify that the package is attached
   if (!"divent" %in% .packages()) {
-    warning("Function arguments cannot be checked because the package divent is not attached. Add CheckArguments=FALSE to suppress this warning or run library('SpatDiv')")
+    warning("Function arguments cannot be checked because the package divent is not attached. Add CheckArguments=FALSE to suppress this warning or run library('divent')")
     return (TRUE)
   }
   # Get the list of arguments of the parent function
@@ -332,6 +460,7 @@ check_divent_args <- function(
       )
     }
   }
+  # correction is checked by match.arg()
   # coverage_estimator is checked by match.arg()
   # estimator is checked by match.arg()
   # distances
@@ -389,6 +518,17 @@ check_divent_args <- function(
       error_message(
         "gamma must be TRUE or FALSE", 
         gamma, 
+        parent_function
+      )
+    }
+  }
+  # global
+  if (!is.na(names(args["global"]))) {
+    global <- eval(expression(global), parent.frame())
+    if (!is.logical(global) | length(global) != 1) {
+      error_message(
+        "global must be TRUE or FALSE", 
+        global, 
         parent_function
       )
     }
@@ -514,6 +654,40 @@ check_divent_args <- function(
         q, 
         parent_function
       )
+    }
+  }
+  # r
+  if (!is.na(names(args["r"]))) {
+    r <- eval(expression(r), parent.frame())
+    if (!is.null(r)) {
+      if (!is.numeric(r) && !is.vector(r)) {
+        error_message(
+          "r must be a numeric vector", 
+          r, 
+          parent_function
+        )
+      }
+      if (length(r) < 2) {
+        error_message(
+          paste("r has length", length(r), "but it should be at least 2)"), 
+          r, 
+          parent_function
+        )
+      }
+      if (r[1] != 0) {
+        error_message(
+          "The first r value must be 0", 
+          r, 
+          parent_function
+        )
+      }
+      if (any(diff(r) <= 0)) {
+        error_message(
+          "Successive values of r must be increasing", 
+          r, 
+          parent_function
+        )
+      }
     }
   }
   # rate
@@ -655,6 +829,18 @@ check_divent_args <- function(
       }
     }
   }
+  if (!is.na(names(args["X"]))) {
+    X <- eval(expression(X), parent.frame())
+    if (!is.null(X)) {
+      if (!inherits(X, "wmppp")) {
+        error_message(
+          "X must be an object of class 'wmppp'", 
+          X, 
+          parent_function
+        )
+      }
+    }
+  }
   
   # All tests passed.
   return (TRUE)
@@ -671,7 +857,7 @@ check_divent_args <- function(
 #' @param sim_dist_matrix A similarity or distance matrix, or a [stats::dist].
 #' @param species_distribution A species distribution, or a named vector.
 #'
-#' @return A similarity matrix that corresponds to the species distribution.
+#' @returns A similarity matrix that corresponds to the species distribution.
 #' @noRd
 #'
 checked_matrix <- function(
@@ -733,7 +919,7 @@ checked_matrix <- function(
 #' 
 #' @param species_distribution An object of class [species_distribution].
 #'
-#' @return A tibble with the estimator used and the estimated entropy.
+#' @returns A tibble with the estimator used and the estimated entropy.
 #' @noRd
 #' 
 ent_gamma_hill <- function(
@@ -865,7 +1051,7 @@ ent_gamma_hill <- function(
 #' See [ent_gamma_hill] for details.
 #' @param species_distribution An object of class [species_distribution].
 #' 
-#' @return A tibble with the estimator used and the estimated entropy.
+#' @returns A tibble with the estimator used and the estimated entropy.
 #' @noRd
 #' 
 ent_gamma_similarity <- function(
@@ -944,7 +1130,7 @@ ent_gamma_similarity <- function(
 #' @param argument The function argument that did not pass the tests.
 #' @param parent_function The name of the function the argument was passed to.
 #'
-#' @return Nothing
+#' @returns Nothing. Used for side effects.
 #' @noRd
 #' 
 error_message <- function(message, argument, parent_function) {
@@ -955,6 +1141,26 @@ error_message <- function(message, argument, parent_function) {
 }
 
 
+#' Extract a column from an fv object
+#' according to an edge-effect correction
+#'
+#' @param fv the function value object, see [spatstat.explore::fv.object].
+#' @param correction the edge-effect correction: 
+#' "isotropic", "translate" or "none"
+#'
+#' @returns a vector with the function values
+#' @noRd
+#'
+correction_fv <- function(fv, correction) {
+  switch(
+    correction,
+    "isotropic" = fv$iso,
+    "translate" = fv$trans,
+    "none" = fv$un
+  )
+}
+
+
 #' Compute Hurlbert's diversity from its entropy
 #' 
 #' Find the effective number of species numerically
@@ -962,7 +1168,7 @@ error_message <- function(message, argument, parent_function) {
 #' @param hurlbert_entropy The entropy.
 #' @param k The order of entropy.
 #'
-#' @return Hurlbert's effective number of species.
+#' @returns Hurlbert's effective number of species.
 #' @noRd
 #' 
 hurlbert_ent2div <- function(hurlbert_entropy, k) {
@@ -993,7 +1199,7 @@ hurlbert_ent2div <- function(hurlbert_entropy, k) {
 #' 
 #' @param x A numeric vector.
 #'
-#' @return `TRUE` if values are integers.
+#' @returns `TRUE` if values are integers.
 #' @noRd
 #'
 is_integer_values <- function(x) {
@@ -1007,7 +1213,7 @@ is_integer_values <- function(x) {
 #' 
 #' Calculate abundances of species and their ancestors along a phylogenetic tree.
 #' 
-#' @return A list of matrices. 
+#' @returns A list of matrices. 
 #' Each matrix contains the abundances of species (in lines) of each community
 #' (in columns) in an interval of the tree.
 #' @noRd
@@ -1051,7 +1257,7 @@ phylo_abd <- function(
 #' @param phylo_abd A list of matrices of abundance (caution: lines are species,
 #' columns are communities).
 #'
-#' @return A vector. Each item is the entropy of a community.
+#' @returns A vector. Each item is the entropy of a community.
 #'  
 #' @noRd
 #'