+++ R/rlseries.R
@@ -14,9 +14,9 @@
 #' @inheritParams check_divent_args
 #' @param n Number of observations.
 #' @param size The size of the distribution.
-#' @param alpha Fisher's \eqn{\alpha}.
+#' @param fisher_alpha Fisher's \eqn{\alpha}.
 #'
-#' @return A numeric vector with the random values drawn from the log-series distribution.
+#' @returns A numeric vector with the random values drawn from the log-series distribution.
 #' @export
 #'
 #' @references
@@ -25,15 +25,15 @@
 #' @examples
 #' # Generate a community made of 10000 individuals with alpha=40
 #' size <- 1E4
-#' alpha <- 40
-#' species_number <- -alpha * log(alpha / (size + alpha))
-#' abundances <- rlseries(species_number, size = 1E5, alpha = 40)
-#' # rCommunity() may be a better choice
-#' autoplot(rcommunity(n = 1, size = 1E4, alpha = 40, distribution = "lseries"))
+#' fisher_alpha <- 40
+#' species_number <- fisher_alpha * log(1 + size / fisher_alpha)
+#' abundances <- rlseries(species_number, size = 1E5, fisher_alpha = 40)
+#' # rcommunity() may be a better choice here
+#' autoplot(rcommunity(1, size = 1E4, fisher_alpha = 40, distribution = "lseries"))
 rlseries <- function(
     n, 
     size, 
-    alpha, 
+    fisher_alpha, 
     show_progress = TRUE, 
     check_arguments = TRUE) {
   # adapted from Dan Lunn, http://www.stats.ox.ac.uk/~dlunn/BS1_05/BS1_Rcode.pdf
@@ -41,7 +41,7 @@ rlseries <- function(
     stop("size is too large to simulate the distribution.")
   }
   # Fisher's x
-  x <- size / (size + alpha)
+  x <- size / (size + fisher_alpha)
   # Draw random numbers between 0 and 1 that are quantiles of the cumulative function
   u <- sort(stats::runif(n))
   # Prepare the corresponding number of abundances